requirements.txt
importlib
numpy
pillow
faiss
ortools
pickle5
scikit-learn
scipy